More refactoring files around a little.
