+++ input/tweet.graphql
@@ -1,0 +1,6 @@
+type Tweet {
+  author: User! @unmodifiable
+  body: String!
+
+  likers: [User!] @hasAndBelongsToMany(as: "liked")
+}
